Fix lots of warnings

+++ OgreMain/include/Compositor/OgreTextureDefinition.h
@@ -105,13 +105,13 @@ namespace Ogre
 
         public:
             TextureTypes::TextureTypes textureType;
-            uint32 width;               // 0 means adapt to target width
-            uint32 height;              // 0 means adapt to target height
-            uint32 depthOrSlices;       // Can never be 0.
-            uint8  numMipmaps;          // 1u to disable mipmaps, 0 to generate until the max
-            bool   bTargetOrientation;  // If true, follows same getOrientationMode as target
-            float  widthFactor;         // multiple of target width to use (if width = 0)
-            float  heightFactor;        // multiple of target height to use (if height = 0)
+            uint32                     width;          // 0 means adapt to target width
+            uint32                     height;         // 0 means adapt to target height
+            uint32                     depthOrSlices;  // Can never be 0.
+            uint8                      numMipmaps;  // 1u to disable mipmaps, 0 to generate until the max
+            bool  bTargetOrientation;               // If true, follows same getOrientationMode as target
+            float widthFactor;                      // multiple of target width to use (if width = 0)
+            float heightFactor;                     // multiple of target height to use (if height = 0)
             /// Use PFG_UNKNOWN to use same format as main target
             PixelFormatGpu format;
             /// "1"  = Disable.
@@ -226,7 +226,7 @@ namespace Ogre
         static inline uint32 encodeTexSource( size_t index, TextureSource textureSource )
         {
             assert( index <= 0x3FFFFFFF && "Texture Source Index out of supported range" );
-            return ( index & 0x3FFFFFFF ) | ( textureSource << 30 );
+            return ( index & 0x3FFFFFFF ) | ( static_cast<uint32>( textureSource ) << 30u );
         }
 
         static void decodeTexSource( uint32 encodedVal, size_t &outIdx, TextureSource &outTexSource );
